requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Data processing
scikit-learn>=1.0.0
PyWavelets>=1.1.0  # For wavelet preprocessing

# Configuration and I/O
pyyaml>=5.4.0
h5py>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0  # For interactive plots

# Progress bars and logging
tqdm>=4.62.0
colorlog>=6.0.0

# Experiment tracking (optional)
wandb>=0.12.0
tensorboard>=2.7.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
isort>=5.10.0

# Optional accelerations
# einops>=0.4.0  # For better tensor operations
# numba>=0.55.0  # For JIT compilation
# apex  # For mixed precision training (NVIDIA GPUs only)

# Optional financial data
# yfinance>=0.1.70  # For downloading financial data
# ta>=0.10.0  # Technical analysis indicators